requirements.txt
absl-py==0.9.0
alembic==1.4.2
astunparse==1.6.3
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.16.0
Keras-Preprocessing==1.1.2
lxml==4.5.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.0
python-dateutil==2.8.1
python-editor==1.0.4
pywin32==228
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
stomp.py==6.1.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
